Seperated out the config implementation
